+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 # Release notes
 
 ## Version 3.2
+* Add support for Visual Studio 2017 RC (issue #47)
 * Support solutions with multiple projects of the same name (issue #39)
   * Show solution folder names to disambiguate
   * Allow multi-project configurations to unambiguously refer to projects by path